task module added

+++ src/main/java/com/main/glory/servicesImpl/PurchaseImpl.java
@@ -3,10 +3,12 @@ package com.main.glory.servicesImpl;
 import com.main.glory.Dao.SupplierRateDao;
 import com.main.glory.Dao.admin.DepartmentDao;
 import com.main.glory.Dao.admin.ReceiverByDao;
+import com.main.glory.Dao.purchase.MaterialPhotosDao;
 import com.main.glory.Dao.purchase.PurchaseDao;
 import com.main.glory.model.admin.ApprovedBy;
 import com.main.glory.model.admin.Department;
 import com.main.glory.model.admin.ReceiverBy;
+import com.main.glory.model.purchase.MaterialPhotos;
 import com.main.glory.model.purchase.Purchase;
 
 import com.main.glory.model.purchase.response.PurchaseResponse;
@@ -28,6 +30,10 @@ public class PurchaseImpl {
 
     @Autowired
     PurchaseDao purchaseDao;
+
+
+    @Autowired
+    MaterialPhotosDao materialPhotosDao;
 
     @Autowired
     SupplierServiceImpl supplierService;
@@ -141,11 +147,22 @@ public class PurchaseImpl {
     }
 
     public void deleteRecordById(Long id) throws Exception {
-        Purchase purchaseExit=purchaseDao.getPurchaseById(id);
-        if(purchaseExit==null)
+        Purchase purchaseExit = purchaseDao.getPurchaseById(id);
+        if (purchaseExit == null)
             throw new Exception("no record found");
 
+        if (purchaseExit.getChecked() == true)
+            throw new Exception("can't delete the record because status is checked");
+
+        List<MaterialPhotos> materialPhotosList = purchaseExit.getMaterialPhotosList();
         purchaseDao.deleteByPurchaseId(id);
+        if (!materialPhotosList.isEmpty())
+        {
+            for(MaterialPhotos materialPhotos:materialPhotosList)
+            {
+                materialPhotosDao.deleteByMaterialId(materialPhotos.getId());
+            }
+        }
     }
